refactor: address thermo-nuclear audit for skill-eval sweep
The skill-eval correctness sweep (PR #485) introduced three issues the
thermo-nuclear audit flagged as presumptive blockers:

1. Duplicated `_scrub_non_finite` between `skill_eval.py` and
   `trajectories/export.py`. Lifted to a shared helper at
   `benchflow/_utils/json_safe.py` exporting `scrub_non_finite` and
   `dumps_finite`. Both call sites now import from the canonical home;
   the duplicates are gone.

2. Hand-rolled `_toml_quote` helper for TOML escaping. Replaced with
   the canonical `tomli_w.dumps()` (already used in
   `benchflow/task/config.py` and already a project dep). Generated
   task.toml is now built as a dict and serialized in one shot, which
   both removes the bespoke escape logic and aligns with the rest of
   the codebase.

3. `skill_eval.py` had grown to 1122 lines (the audit's 1k-line cap is
   a presumptive blocker). Converted to a package at
   `benchflow/skill_eval/` split into:
     - `_core.py` (778 lines) — dataclasses, dataset loading, task
       generation, SkillEvaluator runner
     - `schema.py` (112 lines) — Pydantic models that validate evals.json
     - `gepa_export.py` (195 lines) — GEPA trace export + trajectory loading
     - `__init__.py` (45 lines) — re-exports preserving the public API
   No file in the package exceeds 1000 lines; every public name
   importable from `benchflow.skill_eval` before is still importable.

Tests updated:
- `tests/test_skill_eval_sweep.py` now imports `scrub_non_finite` /
  `dumps_finite` from `benchflow._utils.json_safe`; the now-deleted
  `_toml_quote` unit test is dropped (the round-trip hostile-input test
  still covers the end-to-end contract).
- `tests/trajectories/test_export_nan_handling.py` monkeypatches the
  renamed `scrub_non_finite` symbol on the export module.

All 2360 tests pass (was 2361 — the one removed test was a unit test
for the deleted `_toml_quote` helper). `ruff check` and `ty check`
are clean.
